More speeeeed refactoring - WIP #228

+++ tools/ld-process-efm/Datatypes/f3frame.h
@@ -48,15 +48,15 @@ private:
     bool isSync0;
     bool isSync1;
 
-    qint32 translateEfm(qint32 efmValue);
-    qint32 getBits(uchar *rawData, qint32 bitIndex, qint32 width);
+    qint16 translateEfm(qint16 efmValue);
+    qint16 getBits(uchar *rawData, qint16 bitIndex, qint16 width);
 };
 
 // The following table provides the 10-bit EFM code (padded with leading
 // zeros to 16-bit) corresponding to 0 to 255.  The represented number is
 // given by the position in the array (i.e. position 0 = EFM code for
 // decimal 0 and so on).
-const qint32 efm2numberLUT[256] = {
+const qint16 efm2numberLUT[256] = {
     0x1220, 0x2100, 0x2420, 0x2220, 0x1100, 0x0110, 0x0420, 0x0900, //   8 (7)
     0x1240, 0x2040, 0x2440, 0x2240, 0x1040, 0x0040, 0x0440, 0x0840, //  16
     0x2020, 0x2080, 0x2480, 0x0820, 0x1080, 0x0080, 0x0480, 0x0880, //  24
